Refactor: 메세지 응답 부분 MessageResponseDto로 통일
