+++ src/main/java/com/example/miniprojectdelivery/enums/UserRoleEnum.java
@@ -1,4 +1,4 @@
-package com.example.miniprojectdelivery.model;
+package com.example.miniprojectdelivery.enums;
 
 public enum UserRoleEnum {
