+++ src/main/java/com/example/miniprojectdelivery/model/Address.java
@@ -7,9 +7,7 @@ import lombok.Setter;
 
 @Entity
 @Getter
-@Setter
 @NoArgsConstructor
-@Table(name = "address")
 public class Address {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
