+++ src/main/java/com/example/miniprojectdelivery/model/Mail.java
@@ -4,7 +4,6 @@ import lombok.Getter;
 import lombok.Setter;
 
 @Getter
-@Setter
 public class Mail {
     private String email;
 }
