+++ src/main/java/com/example/miniprojectdelivery/model/Menu.java
@@ -1,34 +1,33 @@
 package com.example.miniprojectdelivery.model;
 
-import com.example.miniprojectdelivery.dto.MenuRequestDto;
+import com.example.miniprojectdelivery.dto.menu.MenuCreateRequestDto;
+import com.example.miniprojectdelivery.dto.menu.MenuUpdateRequestDto;
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import lombok.Setter;
 
 @Entity
 @Getter
-@Setter
 @NoArgsConstructor
-@Table(name = "menu")
 public class Menu {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
+    private String image;  // S3연동예정, 더미필드(임시)
+
+    @Column(nullable = false)
+    private String name;
+
+    @Column(nullable = false)
+    private int cost;
+
     @ManyToOne   // Restaurant에서 @OneToMany 필요함
     @JoinColumn(name = "restaurant_id")
     private Restaurant restaurant;
 
-    @Column(nullable = false)
-    private String image;  // S3연동예정, 더미필드(임시)
-    @Column(nullable = false)
-    private String name;
-    @Column(nullable = false)
-    private int cost;
-
-    public Menu(MenuRequestDto requestDto, Restaurant restaurant) {
+    public Menu(MenuCreateRequestDto requestDto, Restaurant restaurant) {
         this.id = getId();
         this.restaurant = restaurant; // 임의 Restaurant 생성
         this.image = requestDto.getImage();
@@ -36,8 +35,7 @@ public class Menu {
         this.cost = requestDto.getCost();
     }
 
-
-    public void update(MenuRequestDto requestDto) {
+    public void update(MenuUpdateRequestDto requestDto) {
         this.image = requestDto.getImage();
         this.name = requestDto.getName();
         this.cost = requestDto.getCost();
