+++ src/main/java/com/example/miniprojectdelivery/model/TimeStamp.java
@@ -1,9 +1,8 @@
 package com.example.miniprojectdelivery.model;
 
+import jakarta.persistence.Column;
 import java.time.LocalDateTime;
-import javax.persistence.Column;
-import javax.persistence.EntityListeners;
-import javax.persistence.MappedSuperclass;
+import jakarta.persistence.*;
 import lombok.Getter;
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
